+++ tests/Main.hs
@@ -1,7 +1,7 @@
 module Main (main) where
 import qualified Util as T
-import qualified ModificationTime
+import qualified FileTime
 
 main :: IO ()
 main = T.testMain $ \ _t -> do
-  T.isolatedRun _t "ModificationTime" ModificationTime.main
+  T.isolatedRun _t "FileTime" FileTime.main